fixing bugs

+++ src/utils/cloudinary.js
@@ -2,23 +2,23 @@ import {v2 as cloudinary} from 'cloudinary';
 import fs from 'fs';
 
 cloudinary.config({
-    cloud_name:process.env.CLOUDINARY_CLOUD_NAME,
-    api_key:process.env.CLOUDINARY_API_KEY,
-    api_secret:process.env.CLOUDINARY_API_SECRET,
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
 })
 
-const uploadOnCloudinary = async(localFilePath)=>{
+const uploadOnCloudinary = async(localFilePath) => {
     try{
-        const  response = await cloudinary.uploader.upload(localFilePath,{
-            resource_type:"auto",
+        if(!localFilePath) return null
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto",
         })
-        //  file has been uplaoded on cloudinary successfully
-        console.log(response.url);
+        fs.unlinkSync(localFilePath) // delete file after successful upload
         return response
     }catch(error){
-        fs.unlinkSync(localFilePath)
+        fs.unlinkSync(localFilePath) // delete file if upload fails
+        return null
     }
 }
 
-
-export {uploadOnCloudinary}
+export { uploadOnCloudinary }
